Rename: move middleware to its own directory
