+++ .babelrc
@@ -8,7 +8,8 @@
           "@router": ["./router"],
           "@utility": ["./utility"],
           "@model": ["./model"],
-          "@database": ["./database"]
+          "@database": ["./database"],
+          "@middleware": ["./middleware"]
         }
       }
     ]
