+++ .vscode/settings.json
@@ -10,5 +10,6 @@
   "editor.defaultFormatter": "esbenp.prettier-vscode",
   "editor.formatOnSave": true,
   "editor.tabSize": 2,
-  "editor.indentSize": "tabSize"
+  "editor.indentSize": "tabSize",
+  "javascript.updateImportsOnFileMove.enabled": "always"
 }
